Added crash monitor
Sample to demonstrate a simple crash listener, this app
attaches to the dsp and waits for an exception to be signaled,
then it tries to reload the base image to recover the dsp
functionality. The path to the baseimage is currently hardcoded.

It comes with a fault app generator, based on dmm copy sample
it will pass a dummy address to the bridge driver which will
create an MMU fault.

Signed-off-by: Omar Ramirez Luna <omar.ramirez@ti.com>