API: qos: removed hardcoded paths for qosdyn
Removed the registration of QOS DLL from the Bridge
QOS module. Now the users of the QOS have to explicitly
register for the QOS by specifying the path of the DLL
prior to calling QOS APIs. Registration of DLLS is done
by using DSPManager_RegisterObject function call.

Otherwise you can use the Bridge dynreg.out utility to
register for the DLL.

Signed-off-by: Omar Ramirez Luna <omar.ramirez@ti.com>