dynreg: remove invalid token on preprocessor expression
With newest toolchain compilers: arm-none-linux-gnueabi-gcc
(Sourcery G++ Lite 2010q1-202) 4.4.1; an error is thrown saying
that it is not allowed too use the arrow token in a preprocessor
expression, which seems to be correct since this is a conditional
expression depending on this value. Since that branch of code is not
in use, it has been removed.

This fixes: error: token "->" is not valid in preprocessor expressions

Signed-off-by: Omar Ramirez Luna <omar.ramirez@ti.com>