dsp: fix compilation dependencies for exports
Due to the method used for copying dsp code to this repo, an implicit
dependency was introduced in:

dsp/ti/dspbridge/dsp/bridge_product/exports/lib/bpwr.a64P

Which was trying to access a gpio in the first bank, if this bank
was unpowered a warning was seen as long as the dsp was ON, seen after
kernel 3.4-rc1 due to some changes in gpio code. Obviously, this gpio
bank is not used by the dsp for anything. This should fix that bug.

All the exports have been updated to avoid any mismatch while compiling.

This is still 4.55 code, only the binaries from this repo are affected.

Signed-off-by: Omar Ramirez Luna <omar.ramirez@ti.com>